+++ Sources/StreamChat/APIClient/Endpoints/EndpointPath+OfflineRequest.swift
@@ -83,6 +83,7 @@ extension EndpointPath {
              .translateMessage,
              .truncateChannel,
              .unblockUsers,
+             .unmuteChannel,
              .unreadCounts,
              .updateChannel,
              .updateLiveLocation,