+++ Tests/StreamChatTests/WebSocketClient/Events/MemberEvents_Tests.swift
@@ -51,11 +51,7 @@ final class MemberEvents_Tests: XCTestCase {
             eventType: .memberAdded,
             cid: .unique,
             user: .dummy(userId: .unique),
-            memberContainer: .init(
-                member: .dummy(),
-                invite: nil,
-                memberRole: nil
-            ),
+            memberContainer: .init(member: .dummy()),
             createdAt: .unique
         )
 
@@ -74,7 +70,10 @@ final class MemberEvents_Tests: XCTestCase {
         XCTAssertEqual(event.cid, eventPayload.cid)
         XCTAssertEqual(event.user.id, eventPayload.user?.id)
         XCTAssertEqual(event.member.id, eventPayload.memberContainer?.member?.user!.id)
-        XCTAssertEqual(event.member.memberRole, eventPayload.memberContainer?.member?.role)
+        XCTAssertEqual(
+            event.member.memberRole,
+            eventPayload.memberContainer?.member?.channelRole.map(MemberRole.init(rawChannelValue:))
+        )
         XCTAssertEqual(event.createdAt, eventPayload.createdAt)
     }
 
@@ -87,11 +86,7 @@ final class MemberEvents_Tests: XCTestCase {
             eventType: .memberUpdated,
             cid: .unique,
             user: .dummy(userId: .unique),
-            memberContainer: .init(
-                member: .dummy(),
-                invite: nil,
-                memberRole: nil
-            ),
+            memberContainer: .init(member: .dummy()),
             createdAt: .unique
         )
 
@@ -110,7 +105,10 @@ final class MemberEvents_Tests: XCTestCase {
         XCTAssertEqual(event.cid, eventPayload.cid)
         XCTAssertEqual(event.user.id, eventPayload.user?.id)
         XCTAssertEqual(event.member.id, eventPayload.memberContainer?.member?.user!.id)
-        XCTAssertEqual(event.member.memberRole, eventPayload.memberContainer?.member?.role)
+        XCTAssertEqual(
+            event.member.memberRole,
+            eventPayload.memberContainer?.member?.channelRole.map(MemberRole.init(rawChannelValue:))
+        )
         XCTAssertEqual(event.createdAt, eventPayload.createdAt)
     }